rename xero module -> http
